Split unix socket IPC to separate files, add interfaces
